pest-detect-dep/pyproject.toml
[tool.poetry]
name = "pest-detection-api"
version = "1.0.0"
description = "Flask API for Filtered Pest Detection Model"
authors = ["Your Name <your.email@example.com>"]
readme = "README_Pest_Detection.md"
packages = [{include = ".", from = "."}]

[tool.poetry.dependencies]
python = "^3.10"
tensorflow = "2.10.0"
flask = "2.3.3"
pillow = "10.0.1"
numpy = "<2.0"
matplotlib = "3.7.2"
seaborn = "0.12.2"
scikit-learn = "1.3.0"
scipy = "1.11.3"
gunicorn = "21.2.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
